+++ source/device/cpu/op/argmin/argmin_ref.c
@@ -77,7 +77,7 @@ static int ref_argmin_fp32(float* input, int* output, const struct argmin_op_par
     return 0;
 }
 
-static int ref_argmin_uint8(uint8_t* input, int* output, const struct argmin_op_param* param)
+static int ref_argmin_uint8(uint8_t* input, uint8_t* output, const struct argmin_op_param* param)
 {
     uint8_t min_value;
     int min_value_index;
@@ -178,7 +178,7 @@ static int run(struct node_ops* node_ops, struct exec_node* exec_node, struct ex
     if (input_tensor->data_type == TENGINE_DT_FP32)
         ref_argmin_fp32((float*)in_data, (int*)out_data, argmin_op_param);
     else if (input_tensor->data_type == TENGINE_DT_UINT8)
-        ref_argmin_uint8((uint8_t*)in_data, (int*)out_data, argmin_op_param);
+        ref_argmin_uint8((uint8_t*)in_data, (uint8_t*)out_data, argmin_op_param);
 
     return 0;
 }
